KVM headers: Trivially move kvm_pmc/kvm_pmu into a separate header
This allows intel/core.c in perf-events to no longer need kvm_host.h

This is not sufficient for a refactor that allows multiple KVMs,
however.
